+++ NEWS.md
@@ -1,3 +1,7 @@
+## Version `v3.0.16` (02.12.2025)
+### Changes:
+* Consume "System User required" attribute from module bootstrap and request system user token only if the attribute value is true (MODSIDECAR-87)
+
 ## Version `v3.0.15` (12.11.2025)
 ### Changes:
 * Bump quarkus version (MODSIDECAR-154)